refactor to two files
