+++ vyper/venom/passes/__init__.py
@@ -12,6 +12,7 @@ from .fix_calloca import FixCalloca
 from .fix_mem_locations import FixMemLocationsPass
 from .float_allocas import FloatAllocas
 from .function_inliner import FunctionInlinerPass
+from .internal_return_copy_forwarding import InternalReturnCopyForwardingPass
 from .invoke_arg_copy_forwarding import InvokeArgCopyForwardingPass
 from .literals_codesize import ReduceLiteralsCodesize
 from .load_elimination import LoadElimination
@@ -22,6 +23,7 @@ from .memmerging import MemMergePass
 from .memory_copy_elision import MemoryCopyElisionPass
 from .phi_elimination import PhiEliminationPass
 from .readonly_memory_args_analysis import ReadonlyMemoryArgsAnalysisPass
+from .readonly_invoke_arg_copy_forwarding import ReadonlyInvokeArgCopyForwardingPass
 from .remove_unused_variables import RemoveUnusedVariablesPass
 from .revert_to_assert import RevertToAssert
 from .sccp import SCCP
